Make musickit-user-token unit tests CI-safe (no Redis required)
The PR CI job runs 'bun run test:unit' without REDIS_KV_REST_API_URL
/ REDIS_KV_REST_API_TOKEN — those secrets are only available to the
production deploy. The new test-musickit-user-token-sync.test.ts
imported parseStoredToken / MAX_TOKEN_LENGTH from the route file,
which transitively pulled in api/_utils/api-handler.ts →
api/_utils/_rate-limit.ts, and the latter calls createRedis() at
module load. CI crashed with 'Missing Redis configuration' before
the suite could even run.

Mirrors the existing pattern (_musickit-jwt.ts, _cookie.ts, etc.):
pure helpers live in an underscore-prefixed utility module that
the route file imports from. The route scanner skips underscore-
prefixed files, so this doesn't accidentally surface a new
endpoint.

  - api/sync/_musickit-user-token-utils.ts (new): pure parsing /
    constants, zero side effects, zero Redis-side imports.
  - api/sync/musickit-user-token.ts: imports the helpers from the
    utility module + re-exports them so any code that was already
    importing them from the route file keeps compiling.
  - tests/test-musickit-user-token-sync.test.ts: imports from the
    utility module instead of the route.

Verified by reproducing the CI environment locally with 'env -u
REDIS_KV_REST_API_URL -u REDIS_KV_REST_API_TOKEN -u REDIS_URL bun
run test:unit' — all 237 tests now pass without Redis (previously
failed with 'Missing Redis configuration' during module load).
